Initial reorganization of module structure
