+++ src/uint.rs
@@ -36,41 +36,16 @@ use core::ops::{BitAnd, BitOr, BitXor, Shl, Shr, Add, Sub, Mul, Div, Rem};
 use {NonZero, Ord, Greater, Equal, Less, Cmp, Pow};
 use bit::{Bit, B0, B1};
 
-use __private::{Trim, SizeOf, PrivateAnd, PrivateXor, PrivateSub, PrivateCmp, PrivateSizeOf,
+use private::{Trim, SizeOf, PrivateAnd, PrivateXor, PrivateSub, PrivateCmp, PrivateSizeOf,
                 ShiftDiff, PrivateDiv, PrivateDivFirstStep, PrivatePow, BitDiff};
 
-use __private::{TrimOut, SizeOfOut, PrivateAndOut, PrivateXorOut, PrivateSubOut, PrivateCmpOut,
+use private::{TrimOut, SizeOfOut, PrivateAndOut, PrivateXorOut, PrivateSubOut, PrivateCmpOut,
                 PrivateSizeOfOut, PrivatePowOut, BitDiffOut};
 
 use consts::{U0, U1};
 use {Or, Shleft, Shright, Sum, Prod, Add1, Sub1, Square};
 
-/**
-The **marker trait** for compile time unsigned integers.
-
-This trait should not be implemented for anything outside this crate.
-
-# Example
-```rust
-use typenum::consts::U3;
-use typenum::uint::Unsigned;
-
-assert_eq!(U3::to_u32(), 3);
-```
-*/
-pub trait Unsigned {
-    fn to_u8() -> u8;
-    fn to_u16() -> u16;
-    fn to_u32() -> u32;
-    fn to_u64() -> u64;
-    fn to_usize() -> usize;
-
-    fn to_i8() -> i8;
-    fn to_i16() -> i16;
-    fn to_i32() -> i32;
-    fn to_i64() -> i64;
-    fn to_isize() -> isize;
-}
+pub use marker_traits::Unsigned;
 
 /**
 The terminating type for `UInt`; it always comes after the most significant bit. `UTerm`
